fix(packaging): rename Less plugin files to .cjs for ESM compatibility
ng-packagr@21.2.3 adds "type": "module" to publish/package.json,
causing Node.js v22 to treat .js files as ESM. These Less plugins
use CommonJS syntax (module.exports), so rename them to .cjs extension.
