+++ components/style/color/bezierEasing.cjs
