+++ components/style/color/bezierEasing.less
@@ -1,1 +1,1 @@
-@plugin "./bezierEasing";
+@plugin "./bezierEasing.cjs";
