+++ components/style/color/colorPalette.less
@@ -2,4 +2,4 @@
 @import "bezierEasing";
 @import "tinyColor";
 
-@plugin "./colorPalette";
+@plugin "./colorPalette.cjs";
