+++ components/style/color/tinyColor.less
@@ -1,1 +1,1 @@
-@plugin "./tinyColor";
+@plugin "./tinyColor.cjs";
